Split classes into modules
